Update variable storage language
This patch updates the UEFI chapter of the spec, specifically the parts
about UEFI variable storage.  This language utilizes UEFI's existing
Capsule Update on Disk and EFI Configuration table mechanisms, providing
independent mechanisms for exporting variables to the OS and importing
updates from the OS.

The mechanism for exporting variables from the firmware uses a UEFI
Configuration Table rather than requiring the firmware to maintain the
Capsule on Disk, so that a platform with no permanent storage can
implement exporting variables to the OS without having to write to
storage, and a system without permanent storage can implement exporting
to the OS without having to implement importing from Capsule on Disk.
To keep things simple, we do not specify Capsule on Disk as an export
mechanism to the OS.

This text also specifies a de facto coherency protocol.  As a result, in
most cases requirements on the firmware producing the configuration
table are "must" requirements, but they only apply if the export
procedure is implemented.  Likewise, the requirements on the OS to
create a Variable Updates Capsule are mostly "should" requirements, but
the requirements on the contents are "must" requirements.